// sceneforge-nestjs/src/scene/interface/art.interface.ts
import { CrewMember } from './crew-memeber.interface';

export interface Art {
  productionDesigner?: CrewMember[];
  artDirector?: CrewMember[];
  setDecorator?: CrewMember[];
  propMaster?: CrewMember[];
  makeupArtist?: CrewMember[];
  costumeDesigner?: CrewMember[];
  hairStylist?: CrewMember[];
} 